+++ docs/default.nix
@@ -20,7 +20,7 @@
 }:
 mkDerivation {
   pname = "hoppy-docs";
-  version = "0.8.1";
+  version = "0.9.0";
   src = ./.;
   libraryHaskellDepends = [
     base haskell-src hoppy-generator hoppy-runtime