+++ std/default.nix
@@ -19,7 +19,7 @@
 }:
 mkDerivation {
   pname = "hoppy-std";
-  version = "0.8.1";
+  version = "0.9.0";
   src = ./.;
   libraryHaskellDepends = [
     base filepath haskell-src hoppy-generator